+++ exchange-simulator/Models/OrderBookLevel.cs
@@ -2,6 +2,6 @@ namespace exchange_simulator.Models;
 
 public sealed record OrderBookLevel(
     decimal Price,
-    decimal Size
+    long Size
     
     );
